refactor WebSecurityConfig
